refactor: move gAuth logic to google services refactor: move googleSheets logic to google services chore: lint and spacing changes to neo4j
