--- src/config/gAuth.js
@@ -1,5 +1,0 @@
-const { OAuth2Client } = require("google-auth-library");
-
-const client = new OAuth2Client(process.env.GOOGLE_API);
-
-module.exports = { client };
