+++ src/utils/gFormsAnswers.js
@@ -1,23 +1,4 @@
-const { google } = require("googleapis");
-
-const service = google.sheets("v4");
-const credentials = require("../config/CF_credentials.json");
-
-const authClient = new google.auth.JWT(
-    credentials.client_email,
-    null,
-    credentials.private_key.replace(/\\n/g, "\n"),
-    ["https://www.googleapis.com/auth/spreadsheets"]
-);
-
-async function authorize() {
-    const token = await authClient.authorize();
-    authClient.setCredentials(token);
-}
-
-authorize();
-
-async function getAllAnswersQuery(sheet_id) {
+async function getAllAnswersQuery({ service, authClient }, sheet_id) {
     try {
         const res = await service.spreadsheets.values.get({
             auth          : authClient,
@@ -33,7 +14,7 @@ async function getAllAnswersQuery(sheet_id) {
     }
 }
 
-async function getEvaluationQuery(sheet_id, email) {
+async function getEvaluationQuery({ service, authClient }, sheet_id, email) {
     const res = await service.spreadsheets.values.get({
         auth          : authClient,
         spreadsheetId : sheet_id,
